refactor project structure and organize imports
