+++ lib/clone.ts
