+++ lib/query-parser.ts
